判断完全二叉树(queue)
judge the existence of the left child is "lchild and lchild.elem": 1. though the attributes of the left child such as lchild.lchild, lchild.elem, and lchild.rchild are None, the whole lchild is exist as an object. So if only judge lchild, that is True.  2. Why not judge lchild.elem only? Because when the lchild is None, it does not have the attribute elem.